Restructure `server.ts` into a separate `server` directory
