+++ munkey/server/index.ts
@@ -1,0 +1,7 @@
+import CommandServer from "./CommandServer";
+import ShellCommandServer from "./ShellCommandServer";
+
+export {
+    CommandServer,
+    ShellCommandServer,
+};
